Have no idea why it's working build
This is after jumping back and forth between old fixing build and new broken build. Went back to old fixing build and right now all features appear to be working, including custom temps for WU day1 and 2, and +1 +2 and +3 for both WU and OW. Data source verified as switching between OW and WU as expected. Verified with baro trend tag on page1, no H: L: tags on page2, and u before temp on temp1 on page3 for day3